+++ Trabalho-Age-of-Mythology-Retold-Arena/src/Guerreiros/Nordicos/TipoGuerreiroNordicos.java
@@ -13,7 +13,7 @@ import Guerreiros.TipoGuerreiro;
  */
 public abstract class TipoGuerreiroNordicos extends TipoGuerreiro {
 
-    public TipoGuerreiroNordicos(String nome, int idade, float peso) {
+    public TipoGuerreiroNordicos(String nome, int idade, double peso) {
         super(nome, idade, peso);
     }
